pyproject.toml
[project]
name = "flare-ai-kit"
version = "0.1.0"
description = "SDK for building verifiable AI Agents on Flare using Confidential Space"
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [
    { name = "Dinesh Pinto", email = "annual.fallout_0z@icloud.com" }
]
keywords = ["flare", "confidential-space", "blockchain", "ai", "agent", "sdk", "confidential computing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.51.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "structlog>=25.5.0",
    "tenacity>=9.0.0,<10.0.0",
    "web3>=7.14.0",
    "gitpython>=3.1.45",
    "google-adk>=1.19.0",
]

[project.urls]
Homepage = "https://github.com/flare-foundation/flare-ai-kit"
Repository = "https://github.com/flare-foundation/flare-ai-kit"
Releases = "https://github.com/flare-foundation/flare-ai-kit/releases"
Issues = "https://github.com/flare-foundation/flare-ai-kit/issues"

[project.scripts]
flare-ai-kit = "flare_ai_kit.main:start"

[project.optional-dependencies]
# Core functionality groups
ftso = [
    # FTSO functionality uses core dependencies (web3, httpx, pydantic)
    # No additional dependencies required
]
da = [
    # Data Availability layer functionality uses core dependencies
    # No additional dependencies required
]
fassets = [
    # FAssets functionality uses core dependencies
    # No additional dependencies required
]
pdf = [
    "pillow>=11.3.0",
    "pymupdf>=1.26.6",
    "pytesseract>=0.3.13",
    "reportlab>=4.4.5",
    "dulwich>=0.24.10"
]
rag = [
    "qdrant-client==1.15.1",
    "dulwich>=0.24.10"
]
a2a = [
    "fastapi[standard]>=0.118.0",
]
# Additional functionality groups
social = [
    "python-telegram-bot>=22.5",
    "tweepy>=4.16.0",
    "async_lru>=2.0.5",
    "slack-sdk>=3.38.0",
    "discord.py>=2.6.4"
]
tee = [
    "cryptography>=46.0.3",
    "pyjwt>=2.10.1",
    "pyopenssl>=25.3.0"
]
wallet = [
    "cryptography>=46.0.3",
    "eth-account>=0.13.7",
    "pyjwt>=2.10.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.5",
    "settings-doc>=4.3.2",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D212", "COM812", "S105", "D401", "D104", "ANN401", "ISC003", "D107", "FBT001", "FBT002", "FBT003"]

[tool.ruff.lint.extend-per-file-ignores]
"src/flare_ai_kit/agent/adk_agent.py" = ["E501"]
"src/flare_ai_kit/agent/ecosystem_tools_wrapper.py" = ["PLC0415"]
"tests/**/*.py" = ["S", "ARG", "PLR2004", "SLF001", "BLE001", "E501", "T201", "D", "ANN", "F821", "PLC"]
"examples/**/*.py" = ["D", "T201", "BLE001", "INP001", "E501"]
"scripts/**/*.py" = ["D", "T201", "BLE001", "INP001", "E501"]

[tool.pyright]
pythonVersion = "3.12"
include = ["src/"]
typeCheckingMode = "strict"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
